Omnibus of Performance & Testing Improvements (#54)
* Fix broken usethesource repo

The old HTTP repo was inaccessible--it's got an HTTPS URL now.

* Fix security errors on JDK17

JDK17 locked down some of the reflection voodoo the test suite wants to
do. We have to add --add-opens to unlock java.base/java.lang and
java.util.

* project.clj: add a :focus test selector

The normal test suite takes forever to run; this makes it easier to
quickly iterate on a small test.

* Add IGraph.edge(from, to, notFound)

We burn a ton of time in Elle asking for edges in graphs which don't
have them, and each of those requires constructing an expensive
exception. This adds an extra arity to edge, similar to Map.get(), which
returns notFound rather than throwing if an edge does not exist.

* Add IGraph.edge(from, to, notFound)

We burn a ton of time in Elle asking for edges in graphs which don't
have them, and each of those requires constructing an expensive
exception. This adds an extra arity to edge, similar to Map.get(), which
returns notFound rather than throwing if an edge does not exist.

* project.clj: add a :focus test selector

The normal test suite takes forever to run; this makes it easier to
quickly iterate on a small test.

* Speed up Graphs.stronglyConnectedComponents on lazy graphs

The SCC routine called .vertices() twice on its input graph, which for
lazily realized graphs could actually take the majority of our CPU time.
This adds a variable and calls .vertices() only once.

* graph-test: fix wrong arity for edge test

Who's got two thumbs and mistook a slow-running test case for successful
completion? This guy.

* Graphs/maps no longer forget their equality on select/intersect

Map.intersection constructed a new Map without any equality functions,
which meant that equality semantics (and performance!) would abruptly
change when maps and graphs were restricted to subsets of their
keys/vertices.

* Graphs.merge: actually merge graphs this time

Also adds a generative test for digraph and Graphs merge.